+++ pypots/imputation/saits/modules/core.py
@@ -19,7 +19,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from ....modules.transformer import EncoderLayer, PositionalEncoding
+from ....nn.modules.transformer import EncoderLayer, PositionalEncoding
 from ....utils.metrics import calc_mae
