+++ pypots/nn/modules/__init__.py
@@ -1,0 +1,6 @@
+"""
+Frequently-used modules like self-attention modules of vanilla Transformer are put in this package.
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: BSD-3-Clause
