+++ pypots/nn/modules/transformer/auto_encoder.py
